feat(ai-chat-ui): consolidate session notifications on session status
Replace the input-needed notification contribution and the completion
notification in the chat input widget with a single contribution that
reacts to session status transitions. Completion notifications now also
cover background sessions. Delegated sessions no longer notify; their
interactions surface in the parent session's UI.
